SSE Changes, Remote Address Loading
This changes a bit of the SSE functionality, as well as added a helper
function:

- `upgrade` (and, inside, `handle_sse`) now accept `FnOnce` instead of
  just `Fn`.  This was an oversight; it is expected that ownership
  can (and should) be transferred into the closure, since the closure
  will only live for the lifetime of the SSE stream.  The handler for
  the `endpoint` function automatically upgrades the `Fn` to `FnOnce`,
  so no change was needed.
- `stream_heartbeat` was added, which is a utility that sends a
  heartbeat every second.  This allows us to close the stream if/when
  the client stops responding.  Essentially, it polls a stream for one
  second; once the timeout is up, it stops polling, sends an event, and
  then loops.

This also adds functionality to handle multiple header values in the
request/response type directly.  This functionality was always available
(through the `headers()` function), but this allows access directly on
the request/response.  This is used for the new remote address loading
functionality.

The remote address loading functionality takes into account the
complexity that an application may face when trying to determine how to
load a client's IP address, since it is very often dependent on how the
infrastructure around the application itself is setup.  The `remote()`
function is likely going to be removed in favor of this new
`remote_address()` funcntionality.