Merge HasHeaders/HasBody into HttpEntity
This removes the `has_body!`, `has_extensions!`, and `has_headers!`
macros, and replaces the first and last with a `HttpEntity` trait
that implements the same behavior.

This also adjusts the `DataStream` implementation to make checking for
limit exceedence non-destructive to the underlying stream.

It also adds optional support for CBOR and MessagePack entities.