Relax Restriction on Toolchains
This relaxes the toolchain requirements for the under library
by adding a `nightly` configuration during the build script.  This
is because the `#[doc(cfg(...))]` is not available on stable (and
likely will not for some time), and there is no easy way to only
enable it if it exists (as it's not required for normal use; just
documenting - but doc-tests still enable the `doc` cfg, so it
can't be gated behind that).  So if the build script detects it's
on the nightly channel, it enables the `nightly` cfg, and the
`#![feature(doc_cfg)]` is gated behind that.  This allows the
library to be used - and tested - on stable.